requirements.txt
# Django Backend
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0

# LLM Integration
ollama>=0.1.0

# Document Processing (REQUIRED)
PyMuPDF>=1.23.0  # Best PDF text extraction (fitz)
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
python-magic-bin>=0.4.14  # For file type detection on Windows
pdfplumber>=0.10.0  # Table extraction from PDFs
python-pptx>=0.6.21  # PowerPoint processing

# RAG Pipeline
tiktoken>=0.5.0  # Token counting

# Audio Processing
faster-whisper>=0.10.0
pydub>=0.25.1

# Utilities
python-decouple>=3.8
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.66.0
